Improve module augmentation for `link` plugin
