+++ packages/ckeditor5-link/src/index.ts
@@ -14,3 +14,9 @@ export { default as LinkImage } from './linkimage';
 export { default as LinkImageEditing } from './linkimageediting';
 export { default as LinkImageUI } from './linkimageui';
 export { default as AutoLink } from './autolink';
+
+export { LinkConfig } from './linkconfig';
+export { default as LinkCommand } from './linkcommand';
+export { default as UnlinkCommand } from './unlinkcommand';
+
+import './augmentation';
